refactor: share normalizer line parsing
